+++ src/main/java/com/saranaresturantsystem/repositories/ProductStoreQtyRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface ProductStoreQtyRepository extends JpaRepository<ProductStoreQty, Long> {
-    Optional<ProductStoreQty> findByProductIdAndStoreId(Long productId, Integer storeId);
+    Optional<ProductStoreQty> findByProductIdAndStoreId(Long productId, Long storeId);
 }
